requirements.txt
lxml
networkx
pillow